chore: move `util` to `tests/setup`
