+++ examples/puppeteer/package.json
@@ -2,7 +2,7 @@
   "private": true,
   "type": "module",
   "scripts": {
-    "test": "uvu tests",
+    "test": "uvu tests -i setup",
     "dev": "vite",
     "build": "vite build"
   },
