+++ examples/puppeteer/tests/math.js
@@ -1,6 +1,6 @@
 import { suite } from 'uvu';
 import * as assert from 'uvu/assert';
-import withPuppeteer from '../util/withPuppeteer.js'
+import withPuppeteer from './setup/withPuppeteer.js'
 
 const math = withPuppeteer(suite('math'));
 
@@ -13,7 +13,7 @@ math('sum', async ({ page }) => {
 
   value = await page.evaluate(() => typeof window.__MATH__.sum);
   assert.is(value, 'function');
-  
+
   value = await page.evaluate(() => window.__MATH__.sum(1, 2));
   assert.is(value, 3);
 
@@ -29,13 +29,13 @@ math('div', async ({ page }) => {
 
   value = await page.evaluate(() => typeof window.__MATH__.div)
   assert.is(value, 'function');
-  
+
   value = await page.evaluate(() => window.__MATH__.div(1, 2))
   assert.is(value, 0.5);
-  
+
   value = await page.evaluate(() => window.__MATH__.div(-1, -2))
   assert.is(value, 0.5);
-  
+
   value = await page.evaluate(() => window.__MATH__.div(-1, 1))
   assert.is(value, -1);
 });
@@ -45,13 +45,13 @@ math('mod', async ({ page }) => {
 
   value = await page.evaluate(() => typeof window.__MATH__.mod)
   assert.is(value, 'function');
-  
+
   value = await page.evaluate(() => window.__MATH__.mod(1, 2))
   assert.is(value, 1);
-  
+
   value = await page.evaluate(() => window.__MATH__.mod(-3, -2))
   assert.is(value, -1);
-  
+
   value = await page.evaluate(() => window.__MATH__.mod(7, 4))
   assert.is(value, 3);
 });
