+++ examples/puppeteer/tests/utils.js
@@ -1,6 +1,6 @@
 import { suite } from 'uvu';
 import * as assert from 'uvu/assert';
-import withPuppeteer from '../util/withPuppeteer.js'
+import withPuppeteer from './setup/withPuppeteer.js'
 
 const util = withPuppeteer(suite('util'));
 
@@ -13,7 +13,7 @@ util('capitalize', async ({ page }) => {
 
   value = await page.evaluate(() => typeof window.__UTILS__.capitalize);
   assert.is(value, 'function');
-  
+
   value = await page.evaluate(() => window.__UTILS__.capitalize('hello'));
   assert.is(value, 'Hello');
 
@@ -26,13 +26,13 @@ util('dashify', async ({ page }) => {
 
   value = await page.evaluate(() => typeof window.__UTILS__.dashify)
   assert.is(value, 'function');
-  
+
   value = await page.evaluate(() => window.__UTILS__.dashify('fooBar'))
   assert.is(value, 'foo-bar');
-  
+
   value = await page.evaluate(() => window.__UTILS__.dashify('FooBar'))
   assert.is(value, 'foo-bar');
-  
+
   value = await page.evaluate(() => window.__UTILS__.dashify('foobar'))
   assert.is(value, 'foobar');
 });
